+++ src/shared/printer.js
@@ -7,7 +7,7 @@ const chalk = require('chalk');
  * @param {string} message - The message to be logged
  */
 function warn(message) {
-	console.error(`${chalk.bgYellow.black(' WARN: ')} ${message}`);
+	console.error(chalk`{bgAnsi256(214).black.bold WARN:)} ${message}`);
 };
 
 /**
@@ -15,21 +15,21 @@ function warn(message) {
  * @param {string} message - The message to log
  */
 function error(message) {
-	console.error(`${chalk.bgRed.bold.black(' ERR: ')} ${message}`);
+	console.error(chalk`{bgRed.bold.black ERR:} ${message}`);
 };
 
 function errorDetails(message) {
 	if (!global.isVerbose) {
 		return;
 	}
-	console.error(`${chalk.bgRed.bold.black(' ERR: ')} ${message}`);
+	console.error(chalk`{bgRed.bold.black ERR:} ${message}`);
 }
 
 function debug(message) {
 	if (!global.isVerbose) {
 		return;
 	}
-	console.log(`${chalk.magenta('DBG:')} ${chalk.dim(message)}`);
+	console.log(chalk`{ansi256(170) DBG:} {dim ${message}}`);
 }
 
 /**
@@ -41,11 +41,11 @@ function info(message) {
 }
 
 function onDirectoryCreated(directoryPath) {
-	console.log(`Directory ${chalk.grey(directoryPath)} ${chalk.bold.green('created')}`);
+	console.log(chalk`{bold.green Created} directory {ansi256(182) ${directoryPath}}`);
 };
 
 function onFileJobFinished(fileName, destination, template) {
-	let message = `${chalk.green('Created')} ${chalk.ansi256(182)(destination)}/${chalk.bold.ansi256(207)(fileName)}`;
+	let message = `${chalk.green.bold('Created')} ${chalk.ansi256(182)(destination)}/${chalk.bold.ansi256(207)(fileName)}`;
 
 	if (template) {
 		message += ` using template ${chalk.bold.cyan(template)}`